+++ lib/conrad/emitters.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Conrad
+  # :nodoc:
+  module Emitters
+    autoload :Sqs,    'conrad/emitters/sqs'
+    autoload :Stdout, 'conrad/emitters/stdout'
+  end
+end