+++ lib/conrad/formatters.rb
@@ -1,0 +1,6 @@
+module Conrad
+  # :nodoc:
+  module Formatters
+    autoload :JSON, 'conrad/formatters/json'
+  end
+end