+++ lib/conrad/processors.rb
@@ -1,1 +1,8 @@
-Dir['conrad/processors/*.rb'].each { |file| require file }
+module Conrad
+  # :nodoc:
+  module Processors
+    autoload :AddTimestamp, 'conrad/processors/add_timestamp'
+    autoload :AddUUID,      'conrad/processors/add_uuid'
+    autoload :Envelope,     'conrad/processors/envelope'
+  end
+end